refactor: Swagger 관련 컴포넌트 개선
- SwaggerAccessDeniedHandler를 common/security 패키지로 이동
- SwaggerAuthenticationEntryPoint에서 불필요한 setStatus 제거 (ErrorResponse에서 처리)
- 불필요한 ServletException throws 제거
