+++ src/main/java/com/tutorialsejong/courseregistration/common/security/SwaggerAccessDeniedHandler.java
@@ -1,18 +1,18 @@
-package com.tutorialsejong.courseregistration.common.exception;
+package com.tutorialsejong.courseregistration.common.security;
 
-import jakarta.servlet.ServletException;
+import com.tutorialsejong.courseregistration.common.exception.ErrorResponse;
+import com.tutorialsejong.courseregistration.common.exception.GlobalErrorCode;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.web.access.AccessDeniedHandler;
-import java.io.IOException;
 
 public class SwaggerAccessDeniedHandler implements AccessDeniedHandler {
 
     @Override
     public void handle(HttpServletRequest request, HttpServletResponse response,
-                       AccessDeniedException accessDeniedException) throws IOException, ServletException {
-        // ErrorResponse를 SecurityErrorCode 또는 GlobalErrorCode 등 원하는 ErrorCode 를 사용하여 생성
+                       AccessDeniedException accessDeniedException) throws IOException {
         ErrorResponse.from(GlobalErrorCode.HANDLE_ACCESS_DENIED)
                 .writeTo(response);
     }
